fix: transfer news data to external files, simulating API request
